+++ src/main/java/ftc/shift/sample/service/UserService.java
@@ -39,7 +39,7 @@ public class UserService {
      *
      * @param userId - Идентификатор пользователя
      */
-    public UserDtoResponse getUser(Integer userId) {
+    public UserDtoResponse getUser(Long userId) {
         return userMapper.userToDtoResponse(userRepository.getOne(userId));
     }
 
@@ -50,7 +50,7 @@ public class UserService {
      * @param updatedUser - Данные для нового пользователя (имя, тип, дата регистрации)
      * @return Обновленный пользователь
      */
-    public UserDtoResponse updateUser(UserDtoRequest updatedUser, Integer userId) {
+    public UserDtoResponse updateUser(UserDtoRequest updatedUser, Long userId) {
         UserDtoResponse user = userMapper.userToDtoResponse(userRepository.findById(userId)
             .orElseThrow(() -> new DataNotFoundException("User not found")));
         user.setName(updatedUser.getName());
@@ -65,7 +65,7 @@ public class UserService {
      *
      * @param userId - Идентификатор пользователя, которого необходимо удалить
      */
-    public void deleteUser(Integer userId) {
+    public void deleteUser(Long userId) {
         userRepository.delete(userRepository.getOne(userId));
     }
